% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoplight-package.R
\docType{package}
\name{stoplight-package}
\alias{stoplight-package}
\title{stoplight: lightweight assertive programming for dplyr}
\description{
The goal of stoplight is to provide a tidy alternative to the `stopifnot` function. It's equivalent to filtering by the condition and making sure you didn't lose any rows. If no rows are lost, the data pipeline continues, if any row doesn't match the condition, the function returns an error.
}
\author{
\strong{Maintainer}: Anthony Ebert \email{anthonyce@dcm.aau.dk}

}
\keyword{internal}
