% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stoplightt.R
\name{stoplight}
\alias{stoplight}
\title{Tidy version of stop_if_not which respects grouping variables}
\usage{
stoplight(.data, ...)
}
\arguments{
\item{.data}{a data frame}

\item{...}{arguments passed to \link{stopifnot}}
}
\description{
\code{stoplight} allows you to do assertive programming in a tidy manner
}
\examples{

library(stoplight)
library(dplyr)

# The function try() is added to let the package install.
# Don't add the try() function when you use it (unless you want to).

iris |> group_by(Species) |> stoplight(n() == 50)
# Passes

iris |> group_by(Species) |> stoplight(n() != 50) |> try()
# Fails

starwars |> stoplight(n_distinct(name) == n())
# Passes

starwars |> stoplight(if_all(where(is.numeric), ~ .x > 0)) |> try()
# Fails

starwars |> stoplight(if_all(where(is.numeric), ~ .x > 0 | is.na(.x)))
# Passes
}
